+++ src/test/java/org/prebid/server/auction/ExchangeServiceTest.java
@@ -4061,7 +4061,7 @@ public class ExchangeServiceTest extends VertxTest {
         // then
         final Deal expectedDeal = Deal.builder()
                 .id("dealId1")
-                .bidfloor(0.0f)
+                .bidfloor(BigDecimal.valueOf(0.0))
                 .ext(mapper.valueToTree(ExtDeal.of(ExtDealLine.of(null, null, null, null))))
                 .build();
         final BidRequest capturedBidRequest = captureBidRequest();
